Move iters.h to src
